test(pi-agents-tmux): split cwd dirty scan coverage
